Partially fixed fetching HTML
HTML of page is now printed to page console: before the html of the extension's popup was bring printed to the popup's console. Button is essentially dysfunctional now, but I'll try to figure out a way to send a message or something so that the function inside of test.js can be called and the HTML will be sent to the console on a click.