Created smooth movement
Created a specific function that can move the overlay to a particular x and y coordinate smoothly (over the course of .75 seconds). Use 'a' and 'b' for a small demo of the movement. Removed arrow key movement, and mouse movement remains but it moves smoothly to a particular content, so it appears buggy (doesn't follow mouse perfectly)